reducing complexity of useCashItem method
-respective code of item has moved to InventoryHandlerAction for an easier reading. This will help a bit for debugging process.
-Rocks has been checked and working all of them :D
-
