+++ src/handling/MapleServerHandler.java
@@ -61,6 +61,7 @@ import handling.channel.handler.GuildHandler;
 import handling.channel.handler.HiredMerchantHandler;
 import handling.channel.handler.InterServerHandler;
 import handling.channel.handler.InventoryHandler;
+import handling.channel.handler.InventoryHandlerAction;
 import handling.channel.handler.ItemMakerHandler;
 import handling.channel.handler.MobHandler;
 import handling.channel.handler.MonsterCarnivalHandler;
@@ -99,23 +100,25 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
     private boolean cs;
     private final List<String> BlockedIP = new ArrayList<String>();
     private final Map<String, Pair<Long, Byte>> tracker = new ConcurrentHashMap<String, Pair<Long, Byte>>();
-    //Screw locking. Doesn't matter.
-//    private static final ReentrantReadWriteLock IPLoggingLock = new ReentrantReadWriteLock();
+    // Screw locking. Doesn't matter.
+    // private static final ReentrantReadWriteLock IPLoggingLock = new
+    // ReentrantReadWriteLock();
     private static final String nl = System.getProperty("line.separator");
     private static final File loggedIPs = new File("LogIPs.txt");
     private static final HashMap<String, FileWriter> logIPMap = new HashMap<String, FileWriter>();
-    //Note to Zero: Use an enumset. Don't iterate through an array.
-    private static final EnumSet<RecvPacketOpcode> blocked = EnumSet.noneOf(RecvPacketOpcode.class), sBlocked = EnumSet.noneOf(RecvPacketOpcode.class);
+    // Note to Zero: Use an enumset. Don't iterate through an array.
+    private static final EnumSet<RecvPacketOpcode> blocked = EnumSet.noneOf(RecvPacketOpcode.class),
+            sBlocked = EnumSet.noneOf(RecvPacketOpcode.class);
 
     public static void reloadLoggedIPs() {
-//        IPLoggingLock.writeLock().lock();
-//        try {
+        // IPLoggingLock.writeLock().lock();
+        // try {
         for (FileWriter fw : logIPMap.values()) {
             if (fw != null) {
                 try {
                     fw.write("=== Closing Log ===");
                     fw.write(nl);
-                    fw.flush(); //Just in case.
+                    fw.flush(); // Just in case.
                     fw.close();
                 } catch (IOException ex) {
                     System.out.println("Error closing Packet Log.");
@@ -139,11 +142,11 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
             System.out.println("Could not reload packet logged IPs.");
             System.out.println(e);
         }
-//        } finally {
-//            IPLoggingLock.writeLock().unlock();
-//        }
+        // } finally {
+        // IPLoggingLock.writeLock().unlock();
+        // }
     }
-    //Return the Filewriter if the IP is logged. Null otherwise.
+    // Return the Filewriter if the IP is logged. Null otherwise.
 
     private static FileWriter isLoggedIP(IoSession sess) {
         String a = sess.getRemoteAddress().toString();
@@ -163,7 +166,8 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
         }
 
     }
-// <editor-fold defaultstate="collapsed" desc="Packet Log Implementation">
+
+    // <editor-fold defaultstate="collapsed" desc="Packet Log Implementation">
     private static final int Log_Size = 10000, Packet_Log_Size = 25;
     private static final ArrayList<LoggedPacket> Packet_Log = new ArrayList<LoggedPacket>(Log_Size);
     private static final ReentrantReadWriteLock Packet_Log_Lock = new ReentrantReadWriteLock();
@@ -177,21 +181,27 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
             if (Packet_Log.size() == Log_Size) {
                 logged = Packet_Log.remove(0);
             }
-            //This way, we don't create new LoggedPacket objects, we reuse them =]
+            // This way, we don't create new LoggedPacket objects, we reuse them =]
             if (logged == null) {
-                logged = new LoggedPacket(packet, op, io.getRemoteAddress().toString(),
-                        c == null ? -1 : c.getAccID(), FileoutputUtil.CurrentReadable_Time(),
+                logged = new LoggedPacket(packet, op, io.getRemoteAddress().toString(), c == null ? -1 : c.getAccID(),
+                        FileoutputUtil.CurrentReadable_Time(),
                         c == null || c.getAccountName() == null ? "[Null]" : c.getAccountName(),
-                        c == null || c.getPlayer() == null || c.getPlayer().getName() == null ? "[Null]" : c.getPlayer().getName(),
-                        c == null || c.getPlayer() == null || c.getPlayer().getMap() == null ? "[Null]" : String.valueOf(c.getPlayer().getMapId()),
-                        c == null || NPCScriptManager.getInstance().getCM(c) == null ? "[Null]" : String.valueOf(NPCScriptManager.getInstance().getCM(c).getNpc()));
+                        c == null || c.getPlayer() == null || c.getPlayer().getName() == null ? "[Null]"
+                        : c.getPlayer().getName(),
+                        c == null || c.getPlayer() == null || c.getPlayer().getMap() == null ? "[Null]"
+                        : String.valueOf(c.getPlayer().getMapId()),
+                        c == null || NPCScriptManager.getInstance().getCM(c) == null ? "[Null]"
+                        : String.valueOf(NPCScriptManager.getInstance().getCM(c).getNpc()));
             } else {
-                logged.setInfo(packet, op, io.getRemoteAddress().toString(),
-                        c == null ? -1 : c.getAccID(), FileoutputUtil.CurrentReadable_Time(),
+                logged.setInfo(packet, op, io.getRemoteAddress().toString(), c == null ? -1 : c.getAccID(),
+                        FileoutputUtil.CurrentReadable_Time(),
                         c == null || c.getAccountName() == null ? "[Null]" : c.getAccountName(),
-                        c == null || c.getPlayer() == null || c.getPlayer().getName() == null ? "[Null]" : c.getPlayer().getName(),
-                        c == null || c.getPlayer() == null || c.getPlayer().getMap() == null ? "[Null]" : String.valueOf(c.getPlayer().getMapId()),
-                        c == null || NPCScriptManager.getInstance().getCM(c) == null ? "[Null]" : String.valueOf(NPCScriptManager.getInstance().getCM(c).getNpc()));
+                        c == null || c.getPlayer() == null || c.getPlayer().getName() == null ? "[Null]"
+                        : c.getPlayer().getName(),
+                        c == null || c.getPlayer() == null || c.getPlayer().getMap() == null ? "[Null]"
+                        : String.valueOf(c.getPlayer().getMapId()),
+                        c == null || NPCScriptManager.getInstance().getCM(c) == null ? "[Null]"
+                        : String.valueOf(NPCScriptManager.getInstance().getCM(c).getNpc()));
             }
             Packet_Log.add(logged);
         } finally {
@@ -205,11 +215,13 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
         private String ip, accName, accId, chrName, packet, mapId, npcId, op, time;
         private long timestamp;
 
-        public LoggedPacket(String p, String op, String ip, int id, String time, String accName, String chrName, String mapId, String npcId) {
+        public LoggedPacket(String p, String op, String ip, int id, String time, String accName, String chrName,
+                String mapId, String npcId) {
             setInfo(p, op, ip, id, time, accName, chrName, mapId, npcId);
         }
 
-        public final void setInfo(String p, String op, String ip, int id, String time, String accName, String chrName, String mapId, String npcId) {
+        public final void setInfo(String p, String op, String ip, int id, String time, String accName, String chrName,
+                String mapId, String npcId) {
             this.ip = ip;
             this.op = op;
             this.time = time;
@@ -225,7 +237,9 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
         @Override
         public String toString() {
             StringBuilder sb = new StringBuilder();
-            sb.append("[IP: ").append(ip).append("] [").append(accId).append('|').append(accName).append('|').append(chrName).append("] [").append(npcId).append('|').append(mapId).append("] [Time: ").append(timestamp).append("] [").append(time).append(']');
+            sb.append("[IP: ").append(ip).append("] [").append(accId).append('|').append(accName).append('|')
+                    .append(chrName).append("] [").append(npcId).append('|').append(mapId).append("] [Time: ")
+                    .append(timestamp).append("] [").append(time).append(']');
             sb.append(nl);
             sb.append("[Op: ").append(op).append("] [").append(packet).append(']');
             return sb.toString();
@@ -236,7 +250,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
         MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
         try {
             MapleServerHandler mbean = new MapleServerHandler();
-            //The log is a static object, so we can just use this hacky method.
+            // The log is a static object, so we can just use this hacky method.
             mBeanServer.registerMBean(mbean, new ObjectName("handling:type=MapleServerHandler"));
         } catch (Exception e) {
             System.out.println("Error registering PacketLog MBean");
@@ -251,13 +265,16 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
     public void writeLog(boolean crash) {
         Packet_Log_Lock.readLock().lock();
         try {
-            FileWriter fw = new FileWriter(new File(Packet_Log_Output + Packet_Log_Index + (crash ? "_DC.txt" : ".txt")), true);
+            FileWriter fw = new FileWriter(
+                    new File(Packet_Log_Output + Packet_Log_Index + (crash ? "_DC.txt" : ".txt")), true);
             String nl = System.getProperty("line.separator");
             for (LoggedPacket loggedPacket : Packet_Log) {
                 fw.write(loggedPacket.toString());
                 fw.write(nl);
             }
-            final String logString = "Log has been written at " + lastDC + " [" + FileoutputUtil.CurrentReadable_Time() + "] - " + numDC + " have disconnected, within " + (System.currentTimeMillis() - lastDC) + " milliseconds. (" + System.currentTimeMillis() + ")";
+            final String logString = "Log has been written at " + lastDC + " [" + FileoutputUtil.CurrentReadable_Time()
+                    + "] - " + numDC + " have disconnected, within " + (System.currentTimeMillis() - lastDC)
+                    + " milliseconds. (" + System.currentTimeMillis() + ")";
             System.out.println(logString);
             fw.write(logString);
             fw.write(nl);
@@ -281,18 +298,33 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
         ThreadManager.getInstance().newTask(() -> {
             long start = System.currentTimeMillis();
             reloadLoggedIPs();
-            RecvPacketOpcode[] block = new RecvPacketOpcode[]{RecvPacketOpcode.NPC_ACTION, RecvPacketOpcode.MOVE_PLAYER, RecvPacketOpcode.PONG, RecvPacketOpcode.MOVE_PET, RecvPacketOpcode.MOVE_SUMMON, RecvPacketOpcode.MOVE_DRAGON, RecvPacketOpcode.MOVE_LIFE, RecvPacketOpcode.MOVE_ANDROID, RecvPacketOpcode.HEAL_OVER_TIME, RecvPacketOpcode.STRANGE_DATA, RecvPacketOpcode.AUTO_AGGRO, RecvPacketOpcode.CANCEL_DEBUFF, RecvPacketOpcode.MOVE_FAMILIAR};
-            RecvPacketOpcode[] serverBlock = new RecvPacketOpcode[]{RecvPacketOpcode.CHANGE_KEYMAP, RecvPacketOpcode.ITEM_PICKUP, RecvPacketOpcode.PET_LOOT, RecvPacketOpcode.TAKE_DAMAGE, RecvPacketOpcode.FACE_EXPRESSION, RecvPacketOpcode.USE_ITEM, RecvPacketOpcode.CLOSE_RANGE_ATTACK, RecvPacketOpcode.MAGIC_ATTACK, RecvPacketOpcode.RANGED_ATTACK, RecvPacketOpcode.ARAN_COMBO, RecvPacketOpcode.SPECIAL_MOVE, RecvPacketOpcode.GENERAL_CHAT, RecvPacketOpcode.MONSTER_BOMB, RecvPacketOpcode.PASSIVE_ENERGY, RecvPacketOpcode.PET_AUTO_POT, RecvPacketOpcode.USE_CASH_ITEM, RecvPacketOpcode.PARTYCHAT, RecvPacketOpcode.CANCEL_BUFF, RecvPacketOpcode.SKILL_EFFECT, RecvPacketOpcode.CHAR_INFO_REQUEST, RecvPacketOpcode.ALLIANCE_OPERATION, RecvPacketOpcode.AUTO_ASSIGN_AP, RecvPacketOpcode.DISTRIBUTE_AP, RecvPacketOpcode.USE_MAGNIFY_GLASS, RecvPacketOpcode.SPAWN_PET, RecvPacketOpcode.SUMMON_ATTACK, RecvPacketOpcode.ITEM_MOVE, RecvPacketOpcode.PARTY_SEARCH_STOP};
+            RecvPacketOpcode[] block = new RecvPacketOpcode[]{RecvPacketOpcode.NPC_ACTION,
+                RecvPacketOpcode.MOVE_PLAYER, RecvPacketOpcode.PONG, RecvPacketOpcode.MOVE_PET,
+                RecvPacketOpcode.MOVE_SUMMON, RecvPacketOpcode.MOVE_DRAGON, RecvPacketOpcode.MOVE_LIFE,
+                RecvPacketOpcode.MOVE_ANDROID, RecvPacketOpcode.HEAL_OVER_TIME, RecvPacketOpcode.STRANGE_DATA,
+                RecvPacketOpcode.AUTO_AGGRO, RecvPacketOpcode.CANCEL_DEBUFF, RecvPacketOpcode.MOVE_FAMILIAR};
+            RecvPacketOpcode[] serverBlock = new RecvPacketOpcode[]{RecvPacketOpcode.CHANGE_KEYMAP,
+                RecvPacketOpcode.ITEM_PICKUP, RecvPacketOpcode.PET_LOOT, RecvPacketOpcode.TAKE_DAMAGE,
+                RecvPacketOpcode.FACE_EXPRESSION, RecvPacketOpcode.USE_ITEM, RecvPacketOpcode.CLOSE_RANGE_ATTACK,
+                RecvPacketOpcode.MAGIC_ATTACK, RecvPacketOpcode.RANGED_ATTACK, RecvPacketOpcode.ARAN_COMBO,
+                RecvPacketOpcode.SPECIAL_MOVE, RecvPacketOpcode.GENERAL_CHAT, RecvPacketOpcode.MONSTER_BOMB,
+                RecvPacketOpcode.PASSIVE_ENERGY, RecvPacketOpcode.PET_AUTO_POT, RecvPacketOpcode.USE_CASH_ITEM,
+                RecvPacketOpcode.PARTYCHAT, RecvPacketOpcode.CANCEL_BUFF, RecvPacketOpcode.SKILL_EFFECT,
+                RecvPacketOpcode.CHAR_INFO_REQUEST, RecvPacketOpcode.ALLIANCE_OPERATION,
+                RecvPacketOpcode.AUTO_ASSIGN_AP, RecvPacketOpcode.DISTRIBUTE_AP, RecvPacketOpcode.USE_MAGNIFY_GLASS,
+                RecvPacketOpcode.SPAWN_PET, RecvPacketOpcode.SUMMON_ATTACK, RecvPacketOpcode.ITEM_MOVE,
+                RecvPacketOpcode.PARTY_SEARCH_STOP};
             blocked.addAll(Arrays.asList(block));
             sBlocked.addAll(Arrays.asList(serverBlock));
             if (Log_Packets) {
                 for (int i = 1; i <= Packet_Log_Size; i++) {
-                    if (!(new File(Packet_Log_Output + i + ".txt")).exists() && !(new File(Packet_Log_Output + i + "_DC.txt")).exists()) {
+                    if (!(new File(Packet_Log_Output + i + ".txt")).exists()
+                            && !(new File(Packet_Log_Output + i + "_DC.txt")).exists()) {
                         Packet_Log_Index = i;
                         break;
                     }
                 }
-                if (Packet_Log_Index <= 0) { //25+ files, do not log
+                if (Packet_Log_Index <= 0) { // 25+ files, do not log
                     Log_Packets = false;
                 }
             }
@@ -301,11 +333,11 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
             System.out.println("Maple Server loaded in " + (System.currentTimeMillis() - start) + "ms.");
 
         });
-      
+
     }
 
     public MapleServerHandler() {
-        //ONLY FOR THE MBEAN
+        // ONLY FOR THE MBEAN
     }
     // </editor-fold>
 
@@ -316,8 +348,11 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
 
     @Override
     public void exceptionCaught(final IoSession session, final Throwable cause) throws Exception {
-        /*	MapleClient client = (MapleClient) session.getAttribute(MapleClient.CLIENT_KEY);
-        log.error(MapleClient.getLogMessage(client, cause.getMessage()), cause);*/
+        /*
+         * MapleClient client = (MapleClient)
+         * session.getAttribute(MapleClient.CLIENT_KEY);
+         * log.error(MapleClient.getLogMessage(client, cause.getMessage()), cause);
+         */
         // cause.printStackTrace();
     }
 
@@ -325,7 +360,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
     public void sessionOpened(final IoSession session) throws Exception {
         // Start of IP checking
         final String address = session.getRemoteAddress().toString().split(":")[0];
-        System.out.println("\nNew connection from "+session.getRemoteAddress()+" @ "+new Date());
+        System.out.println("\nNew connection from " + session.getRemoteAddress() + " @ " + new Date());
         if (BlockedIP.contains(address)) {
             session.close();
             return;
@@ -376,8 +411,10 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
             }
         }
         LoginServer.removeIPAuth(IP);
-        final byte ivRecv[] = new byte[]{(byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255)};
-        final byte ivSend[] = new byte[]{(byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255)};
+        final byte ivRecv[] = new byte[]{(byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255),
+            (byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255)};
+        final byte ivSend[] = new byte[]{(byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255),
+            (byte) Randomizer.nextInt(255), (byte) Randomizer.nextInt(255)};
 
         final MapleClient client = new MapleClient(
                 new MapleAESOFB(ivSend, (short) (0xFFFF - ServerConstants.MAPLE_VERSION)), // Sent Cypher
@@ -389,7 +426,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
         session.setAttribute(MaplePacketDecoder.DECODER_STATE_KEY, decoderState);
 
         session.write(LoginPacket.getHello(ServerConstants.MAPLE_VERSION, ivSend, ivRecv));
-        //System.out.println("GETHELLO SENT TO " + address);
+        // System.out.println("GETHELLO SENT TO " + address);
         session.setAttribute(MapleClient.CLIENT_KEY, client);
         session.setIdleTime(IdleStatus.READER_IDLE, 60);
         session.setIdleTime(IdleStatus.WRITER_IDLE, 60);
@@ -432,7 +469,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
     @Override
     public void sessionClosed(final IoSession session) throws Exception {
         final MapleClient client = (MapleClient) session.getAttribute(MapleClient.CLIENT_KEY);
-        System.out.println("\nClosing connection from "+session.getRemoteAddress()+" @ "+new Date());
+        System.out.println("\nClosing connection from " + session.getRemoteAddress() + " @ " + new Date());
         if (client != null) {
             byte state = MapleClient.CHANGE_CHANNEL;
             if (Log_Packets && !LoginServer.isShutdown() && !cs && channel > -1) {
@@ -440,17 +477,17 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
             }
             if (state != MapleClient.CHANGE_CHANNEL) {
                 log("Data: " + numDC, "CLOSED", client, session);
-                if (System.currentTimeMillis() - lastDC < 60000) { //within the minute
+                if (System.currentTimeMillis() - lastDC < 60000) { // within the minute
                     numDC++;
-                    if (numDC > 100) { //100+ people have dc'd in minute in channelserver
+                    if (numDC > 100) { // 100+ people have dc'd in minute in channelserver
                         System.out.println("Writing log...");
                         writeLog();
                         numDC = 0;
-                        lastDC = System.currentTimeMillis(); //intentionally place here
+                        lastDC = System.currentTimeMillis(); // intentionally place here
                     }
                 } else {
                     numDC = 0;
-                    lastDC = System.currentTimeMillis(); //intentionally place here
+                    lastDC = System.currentTimeMillis(); // intentionally place here
                 }
             }
             try {
@@ -474,9 +511,10 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
     public void sessionIdle(final IoSession session, final IdleStatus status) throws Exception {
         final MapleClient client = (MapleClient) session.getAttribute(MapleClient.CLIENT_KEY);
 
-        /*	if (client != null && client.getPlayer() != null) {
-        System.out.println("Player "+ client.getPlayer().getName() +" went idle");
-        }*/
+        /*
+         * if (client != null && client.getPlayer() != null) {
+         * System.out.println("Player "+ client.getPlayer().getName() +" went idle"); }
+         */
         if (client != null) {
             client.sendPing();
         }
@@ -499,9 +537,10 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
         final short header_num = slea.readShort();
 
         if (PacketLogging.Log_Packet_Receives) {
-            final StringBuilder sb = new StringBuilder("Client to Server: \n"); 
+            final StringBuilder sb = new StringBuilder("Client to Server: \n");
             sb.append(HexTool.toString((byte[]) message));
-            sb.append(HexTool.toString((byte[]) message)).append("\n").append(HexTool.toStringFromAscii((byte[]) message));
+            sb.append(HexTool.toString((byte[]) message)).append("\n")
+                    .append(HexTool.toStringFromAscii((byte[]) message));
             System.out.println(sb.toString());
         }
         for (final RecvPacketOpcode recv : RecvPacketOpcode.values()) {
@@ -513,24 +552,27 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 }
                 try {
                     if (c.getPlayer() != null && c.isMonitored() && !blocked.contains(recv)) {
-                        FileWriter fw = new FileWriter(new File("MonitorLogs/" + c.getPlayer().getName() + "_log.txt"), true);
-                        fw.write(String.valueOf(recv) + " (" + Integer.toHexString(header_num) + ") Handled: \r\n" + slea.toString() + "\r\n");
+                        FileWriter fw = new FileWriter(new File("MonitorLogs/" + c.getPlayer().getName() + "_log.txt"),
+                                true);
+                        fw.write(String.valueOf(recv) + " (" + Integer.toHexString(header_num) + ") Handled: \r\n"
+                                + slea.toString() + "\r\n");
                         fw.flush();
                         fw.close();
                     }
-                    //no login packets
+                    // no login packets
                     if (Log_Packets && !blocked.contains(recv) && !sBlocked.contains(recv) && (cs || channel > -1)) {
                         log(slea.toString(), recv.toString(), c, session);
                     }
                     handlePacket(recv, slea, c, cs);
-                    //Log after the packet is handle. You'll see why =]
+                    // Log after the packet is handle. You'll see why =]
                     FileWriter fw = isLoggedIP(session);
                     if (fw != null && !blocked.contains(recv)) {
                         if (recv == RecvPacketOpcode.PLAYER_LOGGEDIN && c != null) { // << This is why. Win.
-                            fw.write(">> [AccountName: "
-                                    + (c.getAccountName() == null ? "null" : c.getAccountName()) + "] | [IGN: "
-                                    + (c.getPlayer() == null || c.getPlayer().getName() == null ? "null" : c.getPlayer().getName()) + "] | [Time: "
-                                    + FileoutputUtil.CurrentReadable_Time() + "]");
+                            fw.write(">> [AccountName: " + (c.getAccountName() == null ? "null" : c.getAccountName())
+                                    + "] | [IGN: "
+                                    + (c.getPlayer() == null || c.getPlayer().getName() == null ? "null"
+                                    : c.getPlayer().getName())
+                                    + "] | [Time: " + FileoutputUtil.CurrentReadable_Time() + "]");
                             fw.write(nl);
                         }
                         fw.write("[" + recv.toString() + "]" + slea.toString(true));
@@ -538,20 +580,23 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                         fw.flush();
                     }
                 } catch (NegativeArraySizeException e) {
-                    //swallow, no one cares
+                    // swallow, no one cares
                     if (!ServerConstants.Use_Fixed_IV) {
                         FileoutputUtil.outputFileError(FileoutputUtil.PacketEx_Log, e);
-                        FileoutputUtil.log(FileoutputUtil.PacketEx_Log, "Packet: " + header_num + "\n" + slea.toString(true));
+                        FileoutputUtil.log(FileoutputUtil.PacketEx_Log,
+                                "Packet: " + header_num + "\n" + slea.toString(true));
                     }
                 } catch (ArrayIndexOutOfBoundsException e) {
-                    //swallow, no one cares
+                    // swallow, no one cares
                     if (!ServerConstants.Use_Fixed_IV) {
                         FileoutputUtil.outputFileError(FileoutputUtil.PacketEx_Log, e);
-                        FileoutputUtil.log(FileoutputUtil.PacketEx_Log, "Packet: " + header_num + "\n" + slea.toString(true));
+                        FileoutputUtil.log(FileoutputUtil.PacketEx_Log,
+                                "Packet: " + header_num + "\n" + slea.toString(true));
                     }
                 } catch (Exception e) {
                     FileoutputUtil.outputFileError(FileoutputUtil.PacketEx_Log, e);
-                    FileoutputUtil.log(FileoutputUtil.PacketEx_Log, "Packet: " + header_num + "\n" + slea.toString(true));
+                    FileoutputUtil.log(FileoutputUtil.PacketEx_Log,
+                            "Packet: " + header_num + "\n" + slea.toString(true));
                 }
 
                 return;
@@ -560,13 +605,15 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
 
         if (PacketLogging.Log_Packet_Unhandled_Receives) {
             final StringBuilder sb = new StringBuilder("Unhandled received data Packet: [ " + header_num + " ] \n");
-            sb.append(HexTool.toString((byte[]) message)).append("\n").append(HexTool.toStringFromAscii((byte[]) message));
+            sb.append(HexTool.toString((byte[]) message)).append("\n")
+                    .append(HexTool.toStringFromAscii((byte[]) message));
             System.out.println(sb.toString());
         }
     }
 
-    public static final void handlePacket(final RecvPacketOpcode header, final LittleEndianAccessor slea, final MapleClient c, final boolean cs) throws Exception {
-        //System.out.println(header);
+    public static final void handlePacket(final RecvPacketOpcode header, final LittleEndianAccessor slea,
+            final MapleClient c, final boolean cs) throws Exception {
+        System.out.println(header);
         switch (header) {
             case PONG:
                 c.pongReceived();
@@ -578,11 +625,11 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 CharLoginHandler.login(slea, c);
                 break;
             case SEND_ENCRYPTED:
-                //if (c.isLocalhost()) {
-                //    CharLoginHandler.login(slea, c);
-                //} else {
-                //    c.getSession().write(LoginPacket.getCustomEncryption());
-                //}
+                // if (c.isLocalhost()) {
+                // CharLoginHandler.login(slea, c);
+                // } else {
+                // c.getSession().write(LoginPacket.getCustomEncryption());
+                // }
                 break;
             case CLIENT_START:
             case CLIENT_FAILED:
@@ -598,10 +645,11 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 CharLoginHandler.ServerListRequest(c);
                 break;
             case CLIENT_HELLO:
-                if (slea.readByte() != 8 || slea.readShort() != ServerConstants.MAPLE_VERSION || !String.valueOf(slea.readShort()).equals(ServerConstants.MAPLE_PATCH)) {
+                if (slea.readByte() != 8 || slea.readShort() != ServerConstants.MAPLE_VERSION
+                        || !String.valueOf(slea.readShort()).equals(ServerConstants.MAPLE_PATCH)) {
                     c.getSession().close();
                 }
-                
+
                 break;
             case CHARLIST_REQUEST:
                 CharLoginHandler.CharlistRequest(slea, c);
@@ -635,7 +683,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 CharLoginHandler.Character_WithoutSecondPassword(slea, c, true, true);
                 break;
             case CHAR_SELECT:
-                CharLoginHandler.Character_WithoutSecondPassword(slea, c, true,true);
+                CharLoginHandler.Character_WithoutSecondPassword(slea, c, true, true);
                 break;
             case VIEW_SELECT_PIC:
                 CharLoginHandler.Character_WithSecondPassword(slea, c, true);
@@ -661,14 +709,15 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                     return;
                 }
                 short data_length = slea.readShort();
-                slea.skip(4); // ?				
-                FileoutputUtil.log("ErrorCodes.rtf", "Client sent crashing packet: Type: " + type_str + "; Error code: " + unk + "; Length: " + data_length + "; Packet: " + slea.toString());
+                slea.skip(4); // ?
+                FileoutputUtil.log("ErrorCodes.rtf", "Client sent crashing packet: Type: " + type_str + "; Error code: "
+                        + unk + "; Length: " + data_length + "; Packet: " + slea.toString());
                 break;
             case ENABLE_SPECIAL_CREATION:
                 c.getSession().write(LoginPacket.enableSpecialCreation(c.getAccID(), true));
                 break;
             case RSA_KEY: // Fix this somehow
-                //c.getSession().write(LoginPacket.getLoginAUTH());
+                // c.getSession().write(LoginPacket.getLoginAUTH());
                 // c.getSession().write(LoginPacket.StrangeDATA());
                 break;
             // END OF LOGIN SERVER
@@ -705,7 +754,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 break;
             case ENTER_MTS:
                 InterServerHandler.EnterMTS(c, c.getPlayer());
-                //InterServerHandler.EnterCS(c, c.getPlayer(), true);
+                // InterServerHandler.EnterCS(c, c.getPlayer(), true);
                 break;
             case MOVE_PLAYER:
                 PlayerHandler.MovePlayer(slea, c, c.getPlayer());
@@ -817,7 +866,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 PlayerHandler.CancelChair(slea.readShort(), c, c.getPlayer());
                 break;
             case WHEEL_OF_FORTUNE:
-                break; //whatever
+                break; // whatever
             case USE_ITEMEFFECT:
                 PlayerHandler.UseItemEffect(slea.readInt(), c, c.getPlayer());
                 break;
@@ -945,13 +994,15 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 break;
             case USE_UPGRADE_SCROLL:
                 c.getPlayer().updateTick(slea.readInt());
-                InventoryHandler.UseUpgradeScroll(slea.readShort(), slea.readShort(), slea.readShort(), c, c.getPlayer(), slea.readByte() > 0);
+                InventoryHandlerAction.UseUpgradeScroll(slea.readShort(), slea.readShort(), slea.readShort(), c,
+                        c.getPlayer(), slea.readByte() > 0);
                 break;
             case USE_FLAG_SCROLL:
             case USE_POTENTIAL_SCROLL:
             case USE_EQUIP_SCROLL:
                 c.getPlayer().updateTick(slea.readInt());
-                InventoryHandler.UseUpgradeScroll(slea.readShort(), slea.readShort(), (short) 0, c, c.getPlayer(), slea.readByte() > 0);
+                InventoryHandlerAction.UseUpgradeScroll(slea.readShort(), slea.readShort(), (short) 0, c, c.getPlayer(),
+                        slea.readByte() > 0);
                 break;
             case USE_SUMMON_BAG:
                 InventoryHandler.UseSummonBag(slea, c, c.getPlayer());
@@ -961,7 +1012,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 break;
             case USE_SKILL_BOOK:
                 c.getPlayer().updateTick(slea.readInt());
-                InventoryHandler.UseSkillBook((byte) slea.readShort(), slea.readInt(), c, c.getPlayer());
+                InventoryHandlerAction.UseSkillBook((byte) slea.readShort(), slea.readInt(), c, c.getPlayer());
                 break;
             case USE_CATCH_ITEM:
                 InventoryHandler.UseCatchItem(slea, c, c.getPlayer());
@@ -970,7 +1021,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 InventoryHandler.UseMountFood(slea, c, c.getPlayer());
                 break;
             case REWARD_ITEM:
-                InventoryHandler.UseRewardItem((byte) slea.readShort(), slea.readInt(), c, c.getPlayer());
+                InventoryHandlerAction.UseRewardItem((byte) slea.readShort(), slea.readInt(), c, c.getPlayer());
                 break;
             case HYPNOTIZE_DMG:
                 MobHandler.HypnotizeDmg(slea, c.getPlayer());
@@ -1087,8 +1138,9 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 CashShopOperation.BuyCashItem(slea, c, c.getPlayer());
                 break;
             case COUPON_CODE:
-                //FileoutputUtil.log(FileoutputUtil.PacketEx_Log, "Coupon : \n" + slea.toString(true));
-                //System.out.println(slea.toString());
+                // FileoutputUtil.log(FileoutputUtil.PacketEx_Log, "Coupon : \n" +
+                // slea.toString(true));
+                // System.out.println(slea.toString());
                 CashShopOperation.CouponCode(slea.readMapleAsciiString(), c);
                 CashShopOperation.CouponCode(slea.readMapleAsciiString(), c);
                 CashShopOperation.doCSPackets(c);
@@ -1143,7 +1195,7 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 PetHandler.MovePet(slea, c.getPlayer());
                 break;
             case PET_CHAT:
-                //System.out.println("Pet chat: " + slea.toString());
+                // System.out.println("Pet chat: " + slea.toString());
                 if (slea.available() < 12) {
                     break;
                 }
@@ -1158,11 +1210,12 @@ public class MapleServerHandler extends IoHandlerAdapter implements MapleServerH
                 } else {
                     pet = c.getPlayer().getPet((byte) slea.readInt());
                 }
-                slea.readByte(); //always 0?
+                slea.readByte(); // always 0?
                 if (pet == null) {
                     return;
                 }
-                PetHandler.PetCommand(pet, PetDataFactory.getPetCommand(pet.getPetItemId(), slea.readByte()), c, c.getPlayer());
+                PetHandler.PetCommand(pet, PetDataFactory.getPetCommand(pet.getPetItemId(), slea.readByte()), c,
+                        c.getPlayer());
                 break;
             case PET_FOOD:
                 PetHandler.PetFood(slea, c, c.getPlayer());
